Added tests to guard against RLP applying to HTTPRouteRule with different method
