+++ testsuite/tests/kuadrant/limitador/method/test_route_subset_method.py
@@ -1,4 +1,4 @@
-"""Tests that RLP for a child of HTTPRouteRule doesn't take effect if there isn't a matching HTTPRouteRule"""
+"""Tests that RLP for a HTTPRouteRule doesn't affect the HTTPRoute with same path but different method"""
 
 import pytest
 
@@ -14,6 +14,10 @@ def route(route, backend):
         backend,
         RouteMatch(path=PathMatch(value="/anything", type=MatchType.PATH_PREFIX), method=HTTPMethod.GET),
     )
+    route.add_rule(
+        backend,
+        RouteMatch(path=PathMatch(value="/anything", type=MatchType.PATH_PREFIX), method=HTTPMethod.POST),
+    )
     return route
 
 
@@ -21,16 +25,17 @@ def route(route, backend):
 def rate_limit(rate_limit):
     """Add limit to the policy"""
     selector = RouteSelector(
-        RouteMatch(path=PathMatch(value="/anything/get", type=MatchType.EXACT), method=HTTPMethod.GET)
+        RouteMatch(path=PathMatch(value="/anything", type=MatchType.PATH_PREFIX), method=HTTPMethod.GET)
     )
-    rate_limit.add_limit("subset", [Limit(5, 10)], route_selectors=[selector])
+    rate_limit.add_limit("anything", [Limit(5, 10)], route_selectors=[selector])
     return rate_limit
 
 
-def test_route_rule_child(client):
-    """Tests that RLP for a child of HTTPRoute rule doesn't apply to the parent HTTPRoute rule"""
-    responses = client.get_many("/anything/get", 5)
+def test_route_subset_method(client):
+    """Tests that RLP for a HTTPRouteRule doesn't apply to separate HTTPRouteRule with different method"""
+    responses = client.get_many("/anything", 5)
     assert all(
         r.status_code == 200 for r in responses
     ), f"Rate Limited resource unexpectedly rejected requests {responses}"
-    assert client.get("/anything/get").status_code == 200
+    assert client.get("/anything").status_code == 429
+    assert client.post("/anything").status_code == 200
